Added Shop

+++ src/main/java/fr/bcecb/state/EndGameScreen.java
@@ -24,7 +24,7 @@ public class EndGameScreen extends PopUpScreenState {
         if (Game.instance().getProfile().getRecord(gameType) > this.time) {
             Game.instance().getProfile().setRecord(gameType, this.time);
         }
-        Game.instance().getProfile().setMoneyAmount(Game.instance().getProfile().getMoneyAmount() + this.moneyEarned);
+        Game.instance().getProfile().setMoneyAmount((long) (Game.instance().getProfile().getMoneyAmount() + this.moneyEarned));
         Game.instance().getProfile().save();
     }
 
@@ -44,7 +44,7 @@ public class EndGameScreen extends PopUpScreenState {
         Text allMoneyTitle = new Text(1007, (width / 2f) - 80f, (height / 2f) + 10f, false, "Votre argent :", 0.6f);
 
         Text bestTime = new Text(1008, (width / 2f) + 50f, (height / 2f), false, findBestTime(this.gameType), 0.6f);
-        Text bankroll = new Text(1009, (width / 2f) + 50f, (height / 2f) + 10f, false, intToMoney((int) Game.instance().getProfile().getMoneyAmount()), 0.6f);
+        Text bankroll = new Text(1009, (width / 2f) + 50f, (height / 2f) + 10f, false, intToMoney(Game.instance().getProfile().getMoneyAmount()), 0.6f);
 
         RoundedRectangle lineSeparation2 = new RoundedRectangle(1010, (width / 2f), (height / 2f) + 20f, 180, 1, true, Constants.COLOR_GREY, Float.MAX_VALUE);
